created a funables and teachable modules and a person superclass
