+++ Person.rb
@@ -1,0 +1,7 @@
+class Person
+
+ def initialize(options={})
+    @age = options.fetch(:age, 0)
+    @name = options.fetch(:name, "")
+  end
+end
