+++ apprentice_teacher.rb
@@ -1,22 +1,24 @@
-class ApprenticeTeacher
+require_relative 'person'
+require_relative 'funables'
+require_relative 'teachables'
+
+class ApprenticeTeacher < Person
   attr_reader :age, :salary, :phase, :target_raise
   attr_accessor :name
 
   def initialize(options={})
-    @age = options.fetch(:age, 0)
-    @name = options.fetch(:name, "")
+    super
     @target_raise = 800
     @phase = 3
   end
 
-  def offer_high_five
-    "High five!"
-  end
+  include Funables, Teachables
 
-  def set_phase(num)
-    @phase = num
-    "Cool, I've always wanted to teach phase #{num}!"
-  end
+
+  # def set_phase(num)
+  #   @phase = num
+  #   "Cool, I've always wanted to teach phase #{num}!"
+  # end
 
   def teach_stuff
     response = ""
