+++ teachables.rb
@@ -1,0 +1,6 @@
+module Teachables
+	def set_phase(num)
+		@phase = num
+		"Cool, I've always wanted to teach phase #{num}!"
+	end
+end
